Add flat views
Structured Views works by making TypedArray views for all
intrinsics in a uniform / storage buffer, so

```js
    struct VertexDesc {
        offset: u32,
        stride: u32,
        size: u32,
        padding: u32,
    };

    struct LineInfo {
        triDiv: u32,
        triMul: u32,
        midMod: u32,
        midDiv: u32,
        oddMod: u32,
        triMod: u32,
        pad0: u32,
        pad1: u32,
    };

    struct VSUniforms {
        worldViewProjection: mat4x4<f32>,
        position: VertexDesc,
        lineInfo: LineInfo,
        color: vec4<f32>,
        lightDirection: vec3<f32>,
    };
    @group(0) @binding(0) var<uniform> vsUniforms: VSUniforms;
```

code like the code above makes 15 Float32Array/Uint32Array
at a minimum. If you had an array of structs it would
quickly become 50, 100, lots of views.

I was wondering if, instead of doing that I could just make
one Float32View, one Uint32View, and then you could use
`setFlatView({ ... })` to set everything.

It works, it's just, do I really want 2 ways to do things?
The advantage to this method is less views. This might be
good if you're mapping buffers.

The disadantages are

1. You can only set, not get.

   With stuctured views you can do `console.log(views.lineInfo.midDiv)`.
   Wtih flat views, you really can't. You'd need to lookup the offset
   of lineInfo.midDiv. I could provide that info but it would be a pita
   to use.

2. You can not pass matrices or vecs to math libraries

   Most math JS math libraries take a Float32Array(3, 4, 16) for
   vec3, vec4, mat4. With structured views you can pass the views
   into the library

   ```js
   mat4.perspective(0.5, 2, 1, 100, views.worldViewProjection)
   ```

   With flat you can't. Could try making a new math library
   but then we're back to the previous issue, you'd need offsets.

3. You can not set things in place.

   This is basically the same as the last but, with structured
   views you can do

   ```js
   views.color[0] = 1;
   views.color[1] = 0;
   views.color[2] = 0;
   views.color[3] = 1;
   ```

   With flat views you can only really call `setFlatView` so

   ```js
   setFlatView(view, {color: [0, 1, 2, 3]});
   ```

   which is effectively, `someFloat32Array.set([0, 1, 2, 3], offset)`

   This is probably twice as slow sometimes as you need
   to create an array to pass to `set`.

So, all of that is to say, I'm not sure adding this path is a
good alternative or just gives a choice that's not really needed.

Other considerations, as it is, I can't easily calculate offsets.
Meaning, if you have a complex nested structurs with arrays. With
structured views, each element will have a TypedArray viewing
the correct part of the arraybuffer it's viewing.

With flat though, I'd have to make nearly as giant a tree of offsets
which would kind of self defeating. One of the points of even thinking
about flat view was less memory. But, because I don't have the giant
tree, the code has to walk the data defintions to figure out the
offsets on each `set`.