allow undefined bindgrouplayouts
WebGPU used to require no undefined bindgroups.
In otherwords, if you had @group(0) and @group(5)
you needed to provide bindgrouplayouts for groups 1,2,3,4
that had 0 entries when you called createPipelineLayout.
That was changed so you can pass in undefined.

This PR switched to returning a sparse array so in the example
above you'd get:
`[bindGroupLayoutDescriptor, empty, empty, empty, empty, bindGroupLayoutDescriptor]`
where as before you'd have gotten 6 bindGroupLayoutDescriptors, 4 of them to
descriptors that have 0 entries.